+++ src/ts/dom.ts
@@ -131,6 +131,9 @@ export class DOM {
    * @param handler the handler to execute an operation on an element
    */
   private forEach(handler: (element: HTMLElement) => void): void {
+    if (!this.elements) {
+      return;
+    }
     this.elements.forEach((element) => {
       handler(element);
     });